requirements.txt
google-api-python-client
pandas
python-dotenv
textblob
scikit-learn
nltk
flask
numpy